+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/repository/Credit.kt
@@ -1,4 +1,4 @@
-package com.mayarafernandes.movieplayer.movieList
+package com.mayarafernandes.movieplayer.movieList.repository
 
 data class Credit(val role: String,
                   val name: String)
