+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/repository/MovieMetadata.kt
@@ -1,4 +1,4 @@
-package com.mayarafernandes.movieplayer.movieList
+package com.mayarafernandes.movieplayer.movieList.repository
 
 data class MovieMetadata(val value: String,
                          val name: String)
