+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/repository/ParentalRating.kt
@@ -1,4 +1,4 @@
-package com.mayarafernandes.movieplayer.movieList
+package com.mayarafernandes.movieplayer.movieList.repository
 
 data class ParentalRating(val scheme: String,
                           val rating: String)
