+++ app/src/main/java/com/mayarafernandes/movieplayer/movieList/repository/service/MovieCallbacks.kt
@@ -1,6 +1,6 @@
-package com.mayarafernandes.movieplayer.movieList.service
+package com.mayarafernandes.movieplayer.movieList.repository.service
 
-import com.mayarafernandes.movieplayer.movieList.Movie
+import com.mayarafernandes.movieplayer.movieList.repository.Movie
 
 interface MovieCallbacks {
     fun onSuccess(movieList: List<Movie>)
